Use static Object[][] for argument arrays.
Use static Object[][] for argument arrays.  A table of arguments is created for any length of arguments per thread.  That's 33000+ objects per thread.  But this avoids churn and burn on the GC which translates to better performance.  In my test OpenGL game I notice better and stable FPS.  A ThreadLocal table is used as suggested by Igoldstein so it should be thread safe now.  I also clear the arguments after the native call incase they need to be GCed themselves.