Move callback invocation in c.s.j.p.w.C.util.CallbackProxy into calling thread
Dispatching the invocation of the callback handler into an executor
makes it impossible to fill [out] parameters, as the return has already
happend and [in] parameters can not be savely used if they are not
marshalled to java code because the calling code will free the parameters
after the call.

To prevent deadlocks ComThread is modified to allow COM calls from the
callback by modifying the ComThread helper to only dispatch the COM call
into the ComThread only if the calling thread has not COM already enabled.

Reference counting was modified, so that now on construction of a
ProxyObject the reference count is AddRef'ed once and Released once on
finalization.