Allow calling COM methods/getters requirering hybrid calling (METHOD+PROPERTYGET convention)
It was found, that calling InchesToPoints method from word type library
failed. Research lead to 

https://www.delphitools.info/2013/04/30/gaining-visual-basic-ole-super-powers/
https://msdn.microsoft.com/en-us/library/windows/desktop/ms221486(v=vs.85).aspx

Summary: there are methods in the word typelibrary that require both
PROPERTYGET _and_ METHOD to be set. With only one of these set the call
fails.

The article from delphitools argues, that automation compatible libraries
need to be compatible with VisualBasic which does not distingish methods
and property getters and will set both flags always.

The MSDN article advises this behaviour: "[...] Some languages cannot 
distinguish between retrieving a property and calling a method. In this 
case, you should set the flags DISPATCH_PROPERTYGET and DISPATCH_METHOD.
[...]"

This further support the advised way from delphitools and was implemented.