ignore stdcall on platforms other than ia32 win32
For native libraries, it's a common practice to declare exported
functions and callbacks as stdcall on win32, so on Java side extend
StdCallCallback. However stdcall is only meaningful on IA32 win32, which
makes it difficult to reuse Java code on other platforms like Android.
To solve this, ignore stdcall on all other platforms, not just wince and
win64 (like it was before this patch).