Reintroduce ComThread and move base functions into ObjectFactory
The removal of the ComThread broke one use case of the win32.com.util
packages. It was now necessary to explicitly handle threading and COM
initialization.

The intention was to broaden the possible use cases of the win32.com.util
package.

This changeset tries to reinstantiate the old functionality while keeping
the benefits of the intended change by introducing a new baseclass.