Pass objects into JNI code in addition to raw pointer value (prevent premature GC)
The Memory/Pointer object delegates the native calls to static methods
in the Native class. This causes the Memory/Pointer object to become
prematurely eligable for GC for this case:

----------------------------

Memory pointer = <init>;
<do something and work on Memory>
String result = pointer.getWideString(0)
<do nothing more with Memory>

----------------------------

getWideString(0) calls back into the JVM and allocates a new java string.
In this case the GC can become active and collect the Memory object.
After this it is possible to run the finalizer, that in turn frees the
native memory backing the memory object. The native function now
possibly accesses freed memory.

The native code needs to hold references to the relevant object itself,
this is implemented here.

Changes:

API relevant changes:
- Deprecate Native#getDirectByteBuffer(long, long)
  it is functional, but discouraged to be used and should be removed in
  future

JNI-Layer (Native.java + dispatch.c):
- All set<Type> and get<Type>, write and read methods that took a
  pointer address are modified to take a pointer object, the raw pointer
  value and an offset
- All invoke<ReturnType> methods that took a pointer adress (function
  pointer) are modified to take a function object and the raw function
  pointer value

Build-System/Native:
- the jni api version as bumped from 5.0.0 to 5.1.0 (the major version
  was raised in the 4.3.0 development cycle, so not modified)
- all prebuild native libraries were replaced with the "OUT_OF_DATE"
  marker, where no replacement could be build immediately

getDirectByteBuffer(long, long) is the only API relevant method, allother
methods modified are package private and so are not considered API.
getDirectByteBuffer is overloaded with a package private variant
getDirectByteBuffer(Pointer, long, long, long) that reflects the above
described changes.

Executed tests:

- linux-amd64:   all unittests of jna pass
- linux-arm:     1 unittest fails (unload of jnalib, emulated
                 environment (qemu), slow execution leads to timeout)
- linux-x86:     all unittests of jna pass (32bit chroot on 64bit kernel)
- windows-64bit: 2 tests fail (before and after: DirectTypeMapperTest
                 and NativeTest#testSizeOf (bool should be 1, is 4)
- windows-32bit: 2 tests fails (one less that before the change
                 (testLongStringGeneration), other errors see x64)

Performance:

Check was done with JMH - the comparison was between master and the
proposed pointer_native_direct branch:

java -jar benchmarks-master.jar -i 10 -wi 10 -f 5

The benchmarks initialize a Memory block and access it via getInt and
getString, one with an offset of 0 and on with an offset of 5 bytes.

The testmatrix:

              testInt testIntWithOffset testString testStringWithOffset
linux64-jdk7     108%              108%       100%                 101%
linux64-jdk8     107%              108%       100%                 100%
linux32-jdk8      97%               97%       112%                 112%
win32-jdk8        90%               95%        98%                  98%
win64-jdk8        95%               95%        97%                  97%