Update User32Util#MessageLoopThread to handle error condition of GetMessage
The MSDN advises to check the return value of GetMessage, instead of
blindingly spinning the main loop:

https://msdn.microsoft.com/de-de/library/windows/desktop/ms644936(v=vs.85).aspx

If GetMessage returns -1 this indicates an error and needs to be handled.

By default the MainLoop will exit on an error and log this condition via
java.util.logging. Users can choose to subclass MessageLoopThread to
customize the behavior.