Correct assumption about resolution of VT_DATE
It was interpreted, that the VT_DATA resolution was limited to hour
increments. At the time of this patch the whole way can't be reconstructed,
but tests in a production environment, comparing different COM
implementations (JIntegra and JNA) exposed the different interpreations.

The current implementation follows the advise from:

https://blogs.msdn.microsoft.com/joshpoley/2007/12/19/datetime-formats-and-conversions/

and the definition from (though this description does not show an example handling minutes):

https://msdn.microsoft.com/de-de/library/cc237601.aspx