Fix calling functions with only VarArgs (without fixed parameters)
The change in 50eb820c129d298c26035f98e6a93aed9fc52257 introduces
the detection of the count of the fixed parameters of a VarArgs call.

The logic detecting whether this call was even a VarArgs call was
changed to be true if fixed parameter count is not null. This is
not correct, as the method call could contain only fixed parameters
or a VarArgs call could be done without fixed parameters.

Closes: #763