Do not include ClassPath attribute in jna-platform.jar
The netbeans copylibs task copies the dependend libraries and modifies
the jar MANIFEST.MF to reference the copied libs.

The task is reimplemented in the build.xml to override the base definition
in build-impl.xml, as build-impl.xml comes from the netbeans project template,
future updates could destroy the change if it is not done in build.xml, which
is considered project depended.

Closes: #776