Only analyse binary to determine hardfloat vs. softfloat on arm
The ELFAnalyser analyses the java binary to determine, whether the
hardfloat/softfloat flags are set. This causes issues if a security
manager is in place and limits file accessibility.

The JNA code base needs read access on the /proc/self/exe symlink and
the referenced binary.

This change reduces the cases where the executing binary is read. On
platforms, that don't need it, the detection step is skipped.

Close: #825