Changing Class<T> to Class<T extends Library> for Native.loadLibrary
It was possible to pass a non-library class into this method and get a runtime exception, when the same check can be done at compile-time, so this fix changes it to be done at compile-time.

After erasure, with this change the return type changes from `Object` to `Library`. So as far as backwards compatibility:

1. If you were assigning the result to `Object`, that will still work
2. If you were passing something that wasn't a `Class<? extends Library>`, your code will now not compile, instead of failing at runtime.
3. I can't remember the rules for binary compatibility.

Fixes #822.