Introduce COMInvokeException and prevent a memory leak in case of failure
Only a subset of errors contain exception data. Only calls to
IDispatch#Invoke generate exception details and potentially a "fault
index" for parameters. Therefore these are separated into an extended
COMInvokeException.

Additionally the EXCEPINFO data was not freeed, leading to memory leaks
in in repeated calls.

Add unittest for COMExceptions and correct Exception extraction in Factory