Direct mapping does not support wrapper mappings
For direct mapped methods the java value is directly passed to the C
side. For wrappers this means, that the object pointer is passed and
not the primitive values.

This changeset removes the mappings for the Wrapper classes and makes
it possible to use a type converter to make the call possible, at the
cost of higher calling costs.

Closes: #918