A NULL pointer passed to a IntegerType based direct mapping causes JVM crash
When a IntergetType based type is defined in a direct mapping and a
NullPointer is passed to that call, a method call is tried on that
NullPointer via JNI, resulting in a JVM crash.

This changeset replaces the value with "0" if a NULL pointer is passed
and adds unittests for the possible calling conventions.

Closes: #905