Update W32Service#stopService to be more resilent to delayed status updates
If the service asked to be stopped does not immediatly update its state
in response to a stop command, the stopService method raised an exception.

This change follows a sample on MSDN:

https://msdn.microsoft.com/en-us/library/windows/desktop/ms686335(v=vs.85).aspx

It only uses the status query to determine the final state and not
to see if the services reacted to the stop request.

Closes: #797