Fix SAFEARRAY ByReference in VARIANT
A SAFEARRAY byref that is wrapped in a VARIANT is a **SAFEARRAY. To
model this in JNA a pseudo structure SAFEARRAYByReference is introduced.

This approach is used, so that the referenced SAFEARRAY is read after
a native call.

To make it possible to pass a SAFEARRAYByReference and an IntByReference
through the ProxyObject helper, c.s.j.platform.win32.COM.util.Convert
is modified to reflectivly invoke a matching VARIANT constructor.

This change was manually tested with the GdPicture.NET library using
ProxyObject and COMLateBindingObject based binding.