COM Call through the `com.sun.jna.platform.win32.COM.util.Factory` loose stack trace information
The execution of business method invoked via 
com.sun.jna.platform.win32.COM.util.Factory is dispatched into a
different thread, that has COM initialized. If the execution of one
of these methods raises an Exception, the exception is unwrapped as
far as possible. This in turn looses the call information on the call
site.

This is corrected here by concatenating the two stacktraces. While
not strictly correct, it results in stack traces, that give enough
information to trace them back to their origin.