ACCESS_ACEStructure can't be constructed from pointer
The ACCESS_ACEStructure can't be constructed from pointer because the
super class ACE_HEADER call Structure#read to initialize the values.

Structure#read requires, that array fields are initialized so that
it can determine the correct layout. This has not yet happend when the
ACE_HEADER constructor calls into read and so initialization has to be
taken care of by the read implementation.