Parse Arm EABI section to detect hard-/softfloat usage
This is an update to PR 863: Old toolchains produce binaries without
hard-/softfloat markers. Rasbian is missinng the markers and the oracle
JDK is also affected.

For hardfloat detection now also the Arm EABI section is also considered.

If either the elf hardfloat flag is set or the Arm EABI section indicates
hardfloat the binary is reported as hardfloat.