Check attach result before using resulting JNIEnv pointer in callback code
The success of attaching to the JVM was checked to late in callback.c.
After trying to attach the native thread to the JVM, the JNIEnv pointer
is passed to `get_thread_storage`, which uses it to access the JVM.

At this point it is assumed, that the JNIEnv pointer is valid and can
be used. This is not correct, as the attach can fail and if so the
JNIEnv pointer is not valid.

The pointer is later validated, but to late, so a SEGFAULT occurs.

The result check needs to be moved directly after the attachment
code, to be effective.