Enable building JNA on JDKs newer than 11
Building on JDK 8 is the recommend (and release) way for JNA. It results
in binaries, that are compatible with Java 6.

In theory JNA can be build with JDK 9-11 and target java 6, but that
results in binaries, that match the class format of java 6, but are
linked against methods, that were added/overriden in JDK 9 (ByteBuffer).

So any build with a JDK newer than 9 will not be binary compatible and
thus needs a major version bump.