[GITHUB-1087] Fix wrong calls to Structure#toArray with zero sized arrays
As a performance optimization calls to Collections#toArray where changed
from providing a correctly sized array to using a zero sized array.

See 07f3ce053d947adadf80a20e8acf8caffca8b6cd

The change is invalid if the object toArray is called on is a Structure,
the calls just look the same. This changeset is a partial revert of the
above mentioned commit.

Closes: #1087