On big endian architecture IntegerType based values are incorrectly decoded in the direct invocation path
For IntegerType base values, the member number is read by a direct field
read from C. This results in a jlong. If the target site expects a
datatype smaller than a jlong only parts of that value are read. On
little endian architectures the "correct" part of the long value is
read, but on big endian architectures, unused high bytes are read
instead of the required low bytes.

This can be worked around by bitshifting the value by the size
difference between the target type and the real type.