[GITHUB-1118] Ensure union with floating point and integer member is passed in integer register
System V x86-64 ABI requires, that in a union aggregate, that contains
Integer and Double members, the parameters must be passed in the integer
registers. I.e. in the case where the java side declares double and int
members, the wrong FFI Type would be found, because the doubles size is
larger than the int member, but the wrong parameter passing method would
be used.