Bug fix in WinNT.java
Please refer to Issue #1128 (KEY_ALL_ACCESS value is incorrect in WinNT.java) for details. There is an explanation why it needs to be fixed. 

In short, The KEY_ALL_ACCESS variable has missing parentheses, which gives an incorrect ORed value. It should match with the original winnt.h C header file in Win32.