Only fetch the latest ant version in Travis script
As of 2020-05-13, there are two versions of Apache Ant 1.9.x binaries available for download.  The regexp parsing incorrectly resolves to a string containing both versions, separated by a newline character, and the download fails.   This change limits the grep to only the last line, so the latest 1.9.x version will be downloaded.