Fix ANSI version of Windows API functions
Calling native methods with a java char[] buffer is usually okay. The
buffer will be larger than required for the ANSI API, but the extra
space will not cause any unsolvable problem.
Interperting the returned string in the char[] buffer using the
String(char[]) constructor (or siblings) will apply the wrong encoding.
The char[] buffer needs to be casted to a byte[] buffer in order to
properly get the string value. This conversion is handled inside the
various toString methods of W32StringUtil.

Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>