Fix ANSI version of Windows API functions API BREAK!
Structures containing strings that are passed to native funtions need to
respect the length of TCHAR-arrays depending on A or W function used.
Using a char array in is not the same as char* in C. As a java char is
16bit while in C it's usually 8bit.
Emulate TCHAR-array by using byte array and multiply the element count
with 1 or 2 depending on the current API setting.

Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>