Allow overriding w32.ascii when running test suite
Passing -Dw32.ascii=true or -Dw32.ascii=false on the command line to ANT
will forward the property to the test suite. This allows to run tests on
both the ANSI and the UNICODE versions of the API explicitly.
Not passing the -Dw32.ascii command line option results in the same
behavior as previous versions, i.e. the property is not set.

Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>