Bootstrap dependencies to local maven repository
The ant maven tasks depend on a maven version, that still uses http
to access maven central. This access was deactivated and requests fail.
To work around this problem, the dependencies are specified manually
and downloaded to the local repository. From there the dependency can
be accessed by the "normal" plugin install procedure.