[GH-1313] Move default darwin prefix to darwin-$arch and support darwin as fallback
The darwin platform traditionally used fatbinaries to support multiple
architectures. This works, but is a problem:

a) when apple decides, that only certain architectures may be
   bundled (it was reported, that the presence of x86 caused
   appstore validation to fail)

b) when upstream does not bundle the artifacts, but builds them
   in isolation.

Given, that JNA also hit (b) and needed post processing to merge
binaries, moving darwin to the same scheme as other OSes (<os>-<arch>)
is considered sensible.

While the resource prefix is changed, the old "darwin" prefix is still
considered for extraction and thus old libraries will continue to work.