Split jpms modules into independet artifacts instead of classifier based solution
The jna-platform artifact depends on jna - this leads to problems when
the JPMS artifacts are used, as the pom.xml for the classifier based
artifacts are identical to the base artifacts. This leads to

  <dependency>
    <groupId>net.java.dev.jna</groupId>
    <artifactId>jna-platform</artifactId>
    <version>5.7.0</version>
    <classifier>jpms</classifier>
  </dependency>

depending on

  <dependency>
    <groupId>net.java.dev.jna</groupId>
    <artifactId>jna</artifactId>
    <version>5.7.0</version>
  </dependency>

so the jna-platform JPMS artifacts pulls in the non-JPMS jna artifact.

To solve this, the JPMS artifacts are moved to custom artifact ids. So

  <dependency>
    <groupId>net.java.dev.jna</groupId>
    <artifactId>jna-platform</artifactId>
    <version>5.8.0</version>
    <classifier>jpms</classifier>
  </dependency>

becomes

  <dependency>
    <groupId>net.java.dev.jna</groupId>
    <artifactId>jna-platform-jpms</artifactId>
    <version>5.8.0</version>
  </dependency>