Check if system property native.encoding is set and use that to determine default charset
JNA used the defaultCharset to determine which encoding to use when
converting strings to native char*. The defaultCharset is set from
the system property file.encoding. Up to JDK 17 its value defaulted
to the system default encoding. From JDK 18 onwards its default value
changed to UTF-8.
JDK 18+ exposes the native encoding as the new system property
native.encoding, prior versions don't have that property and will
report NULL for it.
The algorithm is simple: If native.encoding is set, it will be used
else the original implementation of Charset#defaultCharset is used.