Fix crash in direct callbacks on mac OS aarch64
libffi requires the number of fixed arguments to an ffi call to be
correctly provided. For the callback case the conversion routine
supplied the wrong number of arguments for the dispatch into java.

The JNI callbacks have the signature:

NativeType Call<type>Method(JNIEnv *env, jobject obj, jmethodID methodID, ...);

so the first three arguments are fixed and the rest is passed as
varargs.