Load jawt library relative to sun.boot.library.path system on unix OSes
At least Ubuntu builds the JDK with RUNPATH set instead of RPATH. The
two differ in their effect on loading transitive dependencies.

RPATHs effect also covers libraries loaded as transitive dependencies,
while RUNPATH does not. In the case of JNA libjawt is loaded by
libdispatch (the native JNA part), which makes it a transtive load.

The solution is to load the library with the full path based on the
sun.boot.library.path system property, which points to the native
library dirs.