Use Cleaner instead of finalizer in core JNA and JNA-platform
The following uses of finalizers were replaced by a Cleaner or removed
because they are unneeded:

- c.s.j.Memory#finalize
- c.s.j.CallbackReference#finalize
- c.s.j.NativeLibrary#finalize

The cleanup handling is moved to the #close method and #dispose is being
deprecated.

Anyone relying on #dispose being called from the GC has to take action
now! If a subclass needs a cleanup after this object goes out of scope,
it will need to register a cleaner.