Handle arrays in structures with differing size
libffi uses ffi_type structures to describe arrays and structures passed
to native. JNA builds these structures and caches them using the
associated classes as keys. For arrays this leads to the situation where
only a single ffi_type was stored for all sizes of the same base type.
This in turn causes wrong behavior in libffi.