support for Eclipse on Windows on Arm64 platform
Eclipse application for Windows 11 on Arm64 platform uses the following constants at startup:
  OS=win32, ARCH=aarch64
to indicate the OS and system architecture it's running on.
This commit adds a relevant value to the Bundle-NativeCode attribute of the JNA jar's manifest
so that the JNA jar (bundle) can be resolved at runtime by Eclipse's OSGi container,
for an Eclipse application for Windows on Arm64 platform.