Add support for DragonFly BSD x86_64
DragonFly BSD is currently x86_64 only, so don't bother with other
architectures at the moment.

The changes are based on the existing support for other BSDs.

Enhance 'native/Makefile' to include more paths for searching for X11
headers, because the X11 header paths are different on BSDs.  In
addition, add the 'X11INC' variable to allow user to specify extra paths
for X11 headers if really necessary.

Tested on DragonFly 6.5 with OpenJDK 17.0.8 and Ant 1.10.13.